+++ ExilenceNextApp/src/components/expansion-panel/ExpansionPanel.tsx
@@ -1,0 +1,45 @@
+import MuiExpansionPanel from '@material-ui/core/ExpansionPanel';
+import MuiExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
+import MuiExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
+import { withStyles } from '@material-ui/core/styles';
+
+export const ExpansionPanel = withStyles({
+  root: {
+    border: '1px solid rgba(0, 0, 0, .125)',
+    boxShadow: 'none',
+    '&:not(:last-child)': {
+      borderBottom: 0
+    },
+    '&:before': {
+      display: 'none'
+    },
+    '&$expanded': {
+      margin: 'auto'
+    }
+  },
+  expanded: {}
+})(MuiExpansionPanel);
+
+export const ExpansionPanelSummary = withStyles(theme => ({
+  root: {
+    backgroundColor: theme.palette.secondary.main,
+    borderBottom: '1px solid rgba(0, 0, 0, .125)',
+    marginBottom: -1,
+    minHeight: 56,
+    '&$expanded': {
+      minHeight: 56
+    }
+  },
+  content: {
+    '&$expanded': {
+      margin: '12px 0'
+    }
+  },
+  expanded: {}
+}))(MuiExpansionPanelSummary);
+
+export const ExpansionPanelDetails = withStyles(theme => ({
+  root: {
+    padding: theme.spacing(2)
+  }
+}))(MuiExpansionPanelDetails);