+++ ExilenceNextApp/src/components/snapshot-history-chart/SnapshotHistoryChart.tsx
@@ -41,6 +41,9 @@ const SnapshotHistoryChart: React.FC<Props> = ({
   }
 
   const options = {
+    chart: {
+      zoomType: 'x'
+    },
     title: {
       text: ''
     },
@@ -72,12 +75,12 @@ const SnapshotHistoryChart: React.FC<Props> = ({
           ]
         },
         marker: {
-          radius: 2
+          radius: 1
         },
         lineWidth: 2,
         states: {
           hover: {
-            lineWidth: 1
+            lineWidth: 2
           }
         },
         threshold: null