+++ ExilenceNextApp/src/utils/item.utils.ts
@@ -140,12 +140,13 @@ export function getItemVariant(
   explicitMods: string[],
   name: string
 ): string {
-  const watchStoneUsesMod = explicitMods.find(em =>
-    em.includes('uses remaining')
-  );
-
-  if (watchStoneUsesMod) {
-    return watchStoneUsesMod.split(' ')[0];
+  if (explicitMods) {
+    const watchStoneUsesMod = explicitMods.find(em =>
+      em.includes('uses remaining')
+    );
+    if (watchStoneUsesMod) {
+      return watchStoneUsesMod.split(' ')[0];
+    }
   }
 
   if (name === 'Impresence') {